requirements.txt
boto==2.49.0
boto3==1.34.120
botocore==1.34.120
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
